Split 'certificate_chain' module in submodules
