+++ mithril-common/src/certificate_chain/mod.rs
@@ -1,0 +1,9 @@
+//! Tools to retrieve, validate the Certificate Chain created by an aggregator
+
+mod certificate_retriever;
+mod certificate_verifier;
+
+pub use certificate_retriever::{CertificateRetriever, CertificateRetrieverError};
+pub use certificate_verifier::{
+    CertificateVerifier, CertificateVerifierError, MithrilCertificateVerifier,
+};
